requirements.txt
flask>=2.0.0
streamlit>=1.20.0
flask-sqlalchemy>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
cryptography>=40.0.0
requests>=2.28.0
gunicorn>=21.0.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0